requirements.txt
Flask==0.9
Jinja2==2.7
MarkupSafe==0.18
Werkzeug==0.8.3
gevent==0.13.8
gevent-socketio==0.3.5-rc2
greenlet==0.4.1
wsgiref==0.1.2
